Fix bug in function that gets words meeting criteria
The bug was due to the fact that pd.mask and pd.where pass the whole pd.Series
to the callable that serves as a condition, thus the function that checks
accented character didn't work because it was design for analyzing a single
word, not an array. The current decision keeps "complex" logic in the
common function that gets words meeting criteria.